+++ vite.config.js
@@ -1,10 +1,10 @@
-import { defineConfig } from "vite";
+import { defineConfig } from "vite"
 
-export default defineConfig({   
+export default defineConfig({
     root: 'src',
     build: {
         outDir: '../dist',
     },
     publicDir: '../public',
-    
+    envPrefix: 'SKIES_ADSB_',
 })